backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# DSA Integration
girder-client==3.1.20
requests==2.31.0

# Image Processing
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==1.24.2

# Data Handling
pandas==1.5.3

# Caching
diskcache==5.6.1
joblib==1.4.2

# System Metrics (optional - for CPU/memory tracking)
psutil==5.9.8
